requirements.txt
streamlit
pandas
python-dotenv
langchain
langchain-community
langchain-openai
langchain-google-genai
langchain-experimental
faiss-cpu
pypdf
unstructured
tabulate
openai